+++ src/algorithms/search/MazeDistanceHeuristic.java
@@ -9,7 +9,7 @@ public class MazeDistanceHeuristic implements IDistanceHeuristic<MazeState> {
      * @return the euclidean distance between the first and second maze states.
      */
     @Override
-    public int getDistance(MazeState first, MazeState second) {
-        return (int) first.getStatePosition().getDistance(second.getStatePosition());
+    public double getDistance(MazeState first, MazeState second) {
+        return first.getStatePosition().getDistance(second.getStatePosition());
     }
 }
